% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collateOntologies.R
\name{collateOntologies}
\alias{collateOntologies}
\title{Collate ontologies}
\usage{
collateOntologies(paths)
}
\arguments{
\item{paths}{Character vector containing directories generated by Analytical Biosciences,
where each directory contains files pertaining to a single part of a dataset.}
}
\value{
A list of data.frames.
Each data.frame corresponds to a single cell ontology term and contains all cluster names (and their dataset IDs) mapping to that ontology.
}
\description{
Collate cluster names according to their mapped cell ontology,
in order to determine the consistency of ontology mappings across datasets.
}
\author{
Aaron Lun
}
