% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSCE.R
\name{createSCE}
\alias{createSCE}
\title{Create a SingleCellExperiment from a data directory}
\usage{
createSCE(dir, relaxed = FALSE)
}
\arguments{
\item{dir}{String containing a path to a directory of files
generated by Analytical Biosciences.}

\item{relaxed}{Logical scalar indicating whether some of the
sanity checks should be relaxed.}
}
\value{
A \linkS4class{SingleCellExperiment} containing:
\itemize{
\item Experimental data in the \code{\link{assays}}, as obtained from \code{expressionMatrix.rds}.
\item Per-cell metadata in the \code{\link{colData}}, as obtained from \code{ontologyMapping.tsv} and \code{phenotype.tsv}.
\item Per-gene metadata in the \code{\link{rowData}}, as obtained from \code{geneID.tsv}
\item Dimensionality reduction results in \code{\link{reducedDims}}, as obtained from \code{phenotype.tsv}.
\item Study-wide metadta in the \code{\link{metadata}}, as obtained from \code{metadata.tsv}.
}
Any inconsistency in the shape of the data will lead to an error.
}
\description{
Create a \linkS4class{SingleCellExperiment} object containing 
experimental data, sample metadata and general metadata from
a directory of files generated by Analytical Biosciences.
}
\details{
This ingests the data files to create a \linkS4class{SingleCellExperiment}
object that can be quickly queried to check the sanity of various inputs.
It will also serve as the basis of any upload to Genentech systems.

Setting \code{relaxed=TRUE} will not check for consistency between the column names of the matrix and the first field of the column metadata.
This will be set to \code{FALSE} in the future.
}
\examples{
\dontrun{
sce <- createSCE("No_1/part_1")
}
}
\author{
Aaron Lun
}
