% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMTX.R
\name{generateMTX}
\alias{generateMTX}
\title{Generate MTX files according to the RDS files in each dataset.}
\usage{
generateMTX(paths)
}
\arguments{
\item{paths}{Character vector containing directories generated by Analytical Biosciences,
where each directory contains files pertaining to a single part of a dataset.}
}
\value{
A vector of MTX file names that is successfully converted from RDS files.
}
\description{
Generate MTX files according to the RDS files in each dataset.
}
\author{
Analytical Biosciences
}
