% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMarkers.R
\name{loadMarkers}
\alias{loadMarkers}
\title{Load marker sets}
\usage{
loadMarkers(dir)
}
\arguments{
\item{dir}{String containing a path to a directory of files
generated by Analytical Biosciences.}
}
\value{
A \linkS4class{List} of \linkS4class{DataFrame}s where each DataFrame contains ranked marker genes for each cluster in the study.
}
\description{
Load marker sets generated by Analytical Biosciences as a \linkS4class{List} of \linkS4class{DataFrame}s in R.
}
\details{
This ingests the marker list to create a list of ranking tables, which will serve as the basis of any upload to Genentech systems.
If no marker list exists, an empty \linkS4class{List} is returned.
}
\examples{
\dontrun{
sce <- loadMarkers("No_1/part_1")
}

}
\author{
Aaron Lun
}
