% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perStudyReport.R
\name{perStudyReport}
\alias{perStudyReport}
\title{Compile a per-study report}
\usage{
perStudyReport(dir, fname = "report.Rmd")
}
\arguments{
\item{dir}{String containing a path to a directory of files
generated by Analytical Biosciences.}

\item{fname}{String containing the output name of the Rmarkdown report.}
}
\value{
An Rmarkdown report is produced at \code{fname} and compiled with \code{\link{render}}.
}
\description{
Compile a Rmarkdown report summarizing a single study of scRNA-seq data,
from a directory of files prepared by Analytical Biosciences.
}
\details{
The Rmarkdown report simply calls \code{\link{createSCE}} and \code{\link{loadMarkers}}
to check that the various files in \code{dir} can, in fact, be read into R data structures.

It will then give a preview of each of the data fields in these objects,
allowing readers to inspect the contents of the files in an efficient manner.
}
\examples{
\dontrun{
perStudyReport("No_1/part_1")
}
}
\author{
Aaron Lun
}
